Move ShareType to data_models utils directory
