+++ client/lib/app/community/widgets/share/app_share.dart
@@ -10,7 +10,7 @@ import 'package:client/services/services.dart';
 import 'package:client/styles/app_styles.dart';
 import 'package:client/utils/height_constained_text.dart';
 import 'package:data_models/analytics/analytics_entities.dart';
-import 'package:data_models/analytics/share_type.dart';
+import 'package:data_models/utils/share_type.dart';
 
 class AppShareData {
   /// Indicates page path.
